requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
flask-mysql-connector==1.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
numpy==1.26.1
pandas==2.1.1
protobuf==4.21.12
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==3.0.0
